govdoc_checker/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-docx==1.1.2
pydantic==2.8.2
reportlab==4.2.2
requests==2.32.3
